Render list of rockets from API get request
